move system startup code from HasherActorSystem to command implementations (distinction between master and slave
