+++ src/main/scala/com/github/leananeuber/hasher/cli/StartMasterCommand.scala
@@ -2,8 +2,13 @@ package com.github.leananeuber.hasher.cli
 
 import java.io.File
 
+import akka.cluster.Cluster
+import com.github.leananeuber.hasher.{AkkaQuickstart, HasherActorSystem}
 import org.backuity.clist
 
+import scala.concurrent.ExecutionContext.Implicits.global
+import scala.concurrent.duration._
+import scala.language.postfixOps
 
 object StartMasterCommand extends clist.Command(
     name = "master",
@@ -25,6 +30,37 @@ object StartMasterCommand extends clist.Command(
   override def defaultPort: Int = CommonStartCommand.defaultMasterPort
 
   override def run(actorSystemName: String): Unit = {
+    val masterHost = host
+    val masterPort = port
+    val system = HasherActorSystem.actorSystem(actorSystemName, HasherActorSystem.configuration(
+      actorSystemName,
+      masterRole,
+      host,
+      port,
+      masterHost,
+      masterPort
+    ))
+    val cluster = Cluster(system)
 
+    // TODO: handle cluster events
+
+    // TODO: start processing
+    cluster.registerOnMemberUp{
+      // create actors
+      // read `input`
+      // start processing
+    }
+
+    // TODO: remove
+    //------------------------------------------------
+    // run example code on status UP
+    cluster.registerOnMemberUp{
+      AkkaQuickstart.runQuickstartExampleOn(system)
+    }
+    // leave cluster after 2 seconds
+    system.scheduler.scheduleOnce(2 seconds) {
+      cluster.leave(cluster.selfAddress)
+    }
+    //------------------------------------------------
   }
 }
